Fix failure when files do not exist

+++ factor/scripts/remove_synced_data.py
@@ -21,16 +21,18 @@ def main(filename):
 
     """
     # Delete file
-    if os.path.isdir(filename):
-        shutil.rmtree(filename)
-    else:
-        os.remove(filename)
+    if os.path.exists(filename):
+        if os.path.isdir(filename):
+            shutil.rmtree(filename)
+        else:
+            os.remove(filename)
 
     # Next delete the parent directory if empty
-    try:
-        os.rmdir(os.path.dirname(filename))
-    except OSError:
-        pass
+    if os.path.exists(os.path.dirname(filename)):
+        try:
+            os.rmdir(os.path.dirname(filename))
+        except OSError:
+            pass
 
 
 if __name__ == '__main__':
